refactor: move brc partner resources page into sites/brc-analytics (#1637)
